vimrc: use xsel to copy/paste from the X clipboard
I used to install vim-gnome in order to get +clipboard, but I found that, at
least on my current config (Ubuntu 18.04 + byobu), it slows down vim because it
starts a X server for that.

Instead, I now use ,y and ,p to copy to/paste from the X clipboard using xsel
(which therefore needs to be installed).